[ADD] 新增area接口

+++ src/main/java/com/jluzh/sysmanagement/infra/mapper/AreaMapper.java
@@ -14,13 +14,13 @@ import java.util.List;
 public interface AreaMapper {
     List<Area> selectList(Area area);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Area record);
 
     int insertSelective(Area record);
 
-    Area selectByPrimaryKey(Integer id);
+    Area selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(Area record);
